Moved OTACallbacks to ble.h
